revision

+++ prepare/Learning06/Rectangle.cs
@@ -1,11 +1,11 @@
 class Rectangle:Shape{
-    private int _lenght;
-    private int _width;
-    public Rectangle(string color,int lenght, int width):base(color){
+    private double _lenght;
+    private double _width;
+    public Rectangle(string color,double lenght, double width):base(color){
         _lenght=lenght;
         _width=width;
     }
-    public override float GetArea(){
+    public override double GetArea(){
     return _lenght*_width;
 }
 }
